+++ Sources/ParseSwift/API/API+Commands.swift
@@ -389,8 +389,7 @@ internal extension API.Command {
 
         var params: [String: String]?
         if let includeParams = include {
-            let joined = includeParams.joined(separator: ",")
-            params = ["include": joined]
+            params = ["include": "\(includeParams)"]
         }
 
         return API.Command<T, T>(